Fixing issues in TeachingBubble and Callout.
