+++ src/components/Callout/index.ts
@@ -1,3 +1,4 @@
 export * from './Callout';
+export * from './CalloutContent';
 export * from './Callout.Props';
 export * from '../../common/DirectionalHint';
