+++ src/components/TeachingBubble/index.ts
@@ -1,2 +1,3 @@
 export * from './TeachingBubble';
+export * from './TeachingBubbleContent';
 export * from './TeachingBubble.Props';
